fix: load compilerOptions from pylon tsconfig.json
This fixes a issue where the path aliases defined in the `tsconfig.json`
of the pylon were not being loaded when generating the schema.

This was because the `compilerOptions` were defined inside the SchemaBuilder,
and not being loaded from the `tsconfig.json` file.

Now the `tsconfig.json` file is loaded and the pre-defined `compilerOptions`
are only used as a fallback.

Closes #15