fix(builder): use property symbol instead of type property
This fixes the the documentation of type fields (e.g. interfaces) in the
schema parser.
Previously, the return type of the property was used to determine the
description of the field. If a type property used the `Date` type,
the description of the `Date` type was used instead of the actual description
of the field.

This commit changes the behavior to use the property symbol instead of the
return type of the property.

closes #76